Hide prometheus and grafana in the at-home section
I don't want to include them since I think that sortof hurts the datadog
component of this and there is no reason to do that.

But it might come up and if folks want to check it out give them a
reference so we're not giving out other URLs or trying to rememeber
commands offhand.